Fix import, handle new response and extract said response
