+++ resources/js/types/Responses/UpdateSkillsResponse.ts
@@ -1,0 +1,7 @@
+import { LevelUpSkill } from '../Skill';
+import { UserLevels } from '../UserLevels';
+
+export interface UpdateSkillsResponse {
+    new_levels: LevelUpSkill[];
+    user_levels: UserLevels;
+}
